remove CMS_DOCS; view documents in preview
